requirements.txt

Flask==2.3.1
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
Flask-RESTful